enhance pie chart lagend
