+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+53",
+  "version": "0.9.0+54",
   "private": true,
   "type": "module",
   "scripts": {
